Extract MacOSProcess common base (JNA/FFM dedup)
The JNA and FFM macOS OSProcess implementations duplicated their field storage,
all the trivial accessors, the command-line/argument/environment memoization,
thread enumeration, the affinity mask, and the pbi_status state mapping (~243
duplicated lines per Sonar).

Mirroring the BSD/Solaris OSProcess pattern, extract an abstract MacOSProcess
base in oshi-common holding that shared surface, with the macOS state mapping
pulled into a shared stateFromStatus() helper. The native attribute reads stay
in the JNA and FFM subclasses behind hooks: updateAttributes() (proc_pidinfo),
queryArgsAndEnvironment() (KERN_PROCARGS2 sysctl), getSoft/HardOpenFileLimit()
(getrlimit), and queryLogicalProcessorCount() (hw.logicalcpu sysctl).

The os field is typed as the MacOperatingSystem base, so the FFM subclass
constructor widens its parameter accordingly (the OS classes pass 'this', which
already IS-A MacOperatingSystem). Validated by the macOS NativeComparisonTest
(JNA==FFM parity).
