+++ oshi-core/src/main/java/oshi/software/os/mac/MacOSProcessJNA.java
@@ -5,14 +5,6 @@
 package oshi.software.os.mac;
 
 import static oshi.software.os.OSProcess.State.INVALID;
-import static oshi.software.os.OSProcess.State.NEW;
-import static oshi.software.os.OSProcess.State.OTHER;
-import static oshi.software.os.OSProcess.State.RUNNING;
-import static oshi.software.os.OSProcess.State.SLEEPING;
-import static oshi.software.os.OSProcess.State.STOPPED;
-import static oshi.software.os.OSProcess.State.WAITING;
-import static oshi.software.os.OSProcess.State.ZOMBIE;
-import static oshi.util.Memoizer.memoize;
 
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -21,8 +13,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,25 +28,21 @@ import com.sun.jna.platform.unix.LibCAPI.size_t;
 import com.sun.jna.platform.unix.Resource;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.common.mac.ThreadInfo;
 import oshi.jna.Struct.CloseableProcTaskAllInfo;
 import oshi.jna.Struct.CloseableRUsageInfoV2;
 import oshi.jna.Struct.CloseableVnodePathInfo;
 import oshi.jna.platform.mac.SystemB;
-import oshi.software.common.AbstractOSProcess;
-import oshi.software.common.os.mac.MacOSThread;
+import oshi.software.common.os.mac.MacOSProcess;
 import oshi.software.common.os.mac.MacOperatingSystem;
-import oshi.software.os.OSThread;
-import oshi.util.GlobalConfig;
 import oshi.util.ParseUtil;
 import oshi.util.platform.mac.SysctlUtil;
 import oshi.util.tuples.Pair;
 
 /**
- * OSProcess implementation
+ * JNA-backed macOS OSProcess.
  */
 @ThreadSafe
-public class MacOSProcessJNA extends AbstractOSProcess {
+public class MacOSProcessJNA extends MacOSProcess {
 
     private static final Logger LOG = LoggerFactory.getLogger(MacOSProcessJNA.class);
 
@@ -89,96 +75,13 @@ public class MacOSProcessJNA extends AbstractOSProcess {
         TICKS_PER_MS = ticksPerSec / 1000L;
     }
 
-    private static final boolean LOG_MAC_SYSCTL_WARNING = GlobalConfig.get(GlobalConfig.OSHI_OS_MAC_SYSCTL_LOGWARNING,
-            false);
-
-    private static final int MAC_RLIMIT_NOFILE = 8;
-
-    // 64-bit flag
-    private static final int P_LP64 = 0x4;
-    /*
-     * macOS States:
-     */
-    private static final int SSLEEP = 1; // sleeping on high priority
-    private static final int SWAIT = 2; // sleeping on low priority
-    private static final int SRUN = 3; // running
-    private static final int SIDL = 4; // intermediate state in process creation
-    private static final int SZOMB = 5; // intermediate state in process termination
-    private static final int SSTOP = 6; // process being traced
-
-    private int majorVersion;
-    private int minorVersion;
-    private final MacOperatingSystem os;
-
-    private Supplier<String> commandLine = memoize(this::queryCommandLine);
-    private Supplier<Pair<List<String>, Map<String, String>>> argsEnviron = memoize(this::queryArgsAndEnvironment);
-
-    private String name = "";
-    private String path = "";
-    private String currentWorkingDirectory;
-    private String user;
-    private String userID;
-    private String group;
-    private String groupID;
-    private State state = INVALID;
-    private int parentProcessID;
-    private int threadCount;
-    private int priority;
-    private long virtualSize;
-    private long residentSetSize;
-    private long memoryFootprint;
-    private long kernelTime;
-    private long userTime;
-    private long startTime;
-    private long upTime;
-    private long bytesRead;
-    private long bytesWritten;
-    private long openFiles;
-    private int bitness;
-    private long minorFaults;
-    private long majorFaults;
-    private long contextSwitches;
-    private long voluntaryContextSwitches;
-    private long involuntaryContextSwitches;
-
     public MacOSProcessJNA(int pid, int major, int minor, MacOperatingSystem os) {
-        super(pid);
-        this.majorVersion = major;
-        this.minorVersion = minor;
-        this.os = os;
+        super(pid, major, minor, os);
         updateAttributes();
     }
 
     @Override
-    public String getName() {
-        return this.name;
-    }
-
-    @Override
-    public String getPath() {
-        return this.path;
-    }
-
-    @Override
-    public String getCommandLine() {
-        return this.commandLine.get();
-    }
-
-    private String queryCommandLine() {
-        return String.join(" ", getArguments());
-    }
-
-    @Override
-    public List<String> getArguments() {
-        return argsEnviron.get().getA();
-    }
-
-    @Override
-    public Map<String, String> getEnvironmentVariables() {
-        return argsEnviron.get().getB();
-    }
-
-    private Pair<List<String>, Map<String, String>> queryArgsAndEnvironment() {
+    protected Pair<List<String>, Map<String, String>> queryArgsAndEnvironment() {
         int pid = getProcessID();
         // Set up return objects
         List<String> args = new ArrayList<>();
@@ -248,109 +151,8 @@ public class MacOSProcessJNA extends AbstractOSProcess {
     }
 
     @Override
-    public String getCurrentWorkingDirectory() {
-        return this.currentWorkingDirectory;
-    }
-
-    @Override
-    public String getUser() {
-        return this.user;
-    }
-
-    @Override
-    public String getUserID() {
-        return this.userID;
-    }
-
-    @Override
-    public String getGroup() {
-        return this.group;
-    }
-
-    @Override
-    public String getGroupID() {
-        return this.groupID;
-    }
-
-    @Override
-    public State getState() {
-        return this.state;
-    }
-
-    @Override
-    public int getParentProcessID() {
-        return this.parentProcessID;
-    }
-
-    @Override
-    public int getThreadCount() {
-        return this.threadCount;
-    }
-
-    @Override
-    public List<OSThread> getThreadDetails() {
-        long now = System.currentTimeMillis();
-        return ThreadInfo.queryTaskThreads(getProcessID()).stream().parallel().map(stat -> {
-            // For long running threads the start time calculation can overestimate
-            long start = Math.max(now - stat.getUpTime(), getStartTime());
-            return new MacOSThread(getProcessID(), stat.getThreadId(), stat.getState(), stat.getSystemTime(),
-                    stat.getUserTime(), start, now - start, stat.getPriority());
-        }).collect(Collectors.toList());
-    }
-
-    @Override
-    public int getPriority() {
-        return this.priority;
-    }
-
-    @Override
-    public long getVirtualSize() {
-        return this.virtualSize;
-    }
-
-    @Override
-    public long getResidentMemory() {
-        return this.residentSetSize;
-    }
-
-    @Override
-    public long getPrivateResidentMemory() {
-        return this.memoryFootprint;
-    }
-
-    @Override
-    public long getKernelTime() {
-        return this.kernelTime;
-    }
-
-    @Override
-    public long getUserTime() {
-        return this.userTime;
-    }
-
-    @Override
-    public long getUpTime() {
-        return this.upTime;
-    }
-
-    @Override
-    public long getStartTime() {
-        return this.startTime;
-    }
-
-    @Override
-    public long getBytesRead() {
-        return this.bytesRead;
-    }
-
-    @Override
-    public long getBytesWritten() {
-        return this.bytesWritten;
-    }
-
-    @Override
-    public long getOpenFiles() {
-        return this.openFiles;
+    protected int queryLogicalProcessorCount() {
+        return SysctlUtil.sysctl("hw.logicalcpu", 1);
     }
 
     @Override
@@ -373,43 +175,6 @@ public class MacOSProcessJNA extends AbstractOSProcess {
         } else {
             return -1L; // not supported
         }
-    }
-
-    @Override
-    public int getBitness() {
-        return this.bitness;
-    }
-
-    @Override
-    public long getAffinityMask() {
-        // macOS doesn't do affinity. Return a bitmask of the current processors.
-        int logicalProcessorCount = SysctlUtil.sysctl("hw.logicalcpu", 1);
-        return logicalProcessorCount < 64 ? (1L << logicalProcessorCount) - 1 : -1L;
-    }
-
-    @Override
-    public long getMinorFaults() {
-        return this.minorFaults;
-    }
-
-    @Override
-    public long getMajorFaults() {
-        return this.majorFaults;
-    }
-
-    @Override
-    public long getContextSwitches() {
-        return this.contextSwitches;
-    }
-
-    @Override
-    public long getVoluntaryContextSwitches() {
-        return this.voluntaryContextSwitches;
-    }
-
-    @Override
-    public long getInvoluntaryContextSwitches() {
-        return this.involuntaryContextSwitches;
     }
 
     @Override
@@ -436,29 +201,7 @@ public class MacOSProcessJNA extends AbstractOSProcess {
                 this.name = Native.toString(taskAllInfo.pbsd.pbi_comm, StandardCharsets.UTF_8);
             }
 
-            switch (taskAllInfo.pbsd.pbi_status) {
-                case SSLEEP:
-                    this.state = SLEEPING;
-                    break;
-                case SWAIT:
-                    this.state = WAITING;
-                    break;
-                case SRUN:
-                    this.state = RUNNING;
-                    break;
-                case SIDL:
-                    this.state = NEW;
-                    break;
-                case SZOMB:
-                    this.state = ZOMBIE;
-                    break;
-                case SSTOP:
-                    this.state = STOPPED;
-                    break;
-                default:
-                    this.state = OTHER;
-                    break;
-            }
+            this.state = stateFromStatus(taskAllInfo.pbsd.pbi_status);
             this.parentProcessID = taskAllInfo.pbsd.pbi_ppid;
             this.userID = Integer.toString(taskAllInfo.pbsd.pbi_uid);
             Passwd pwuid = SystemB.INSTANCE.getpwuid(taskAllInfo.pbsd.pbi_uid);
